requirements.txt
fastapi
uvicorn
sqlalchemy
netCDF4
pydantic
numpy
psycopg2-binary
requests
geopy shapely
geopy
pandas
openpyxl
